Published version 4.5.0.
It includes the fix for the bug reported by Allen and investigated by Angel (rev. 8261eda2e8b7fb9eb86ce79b16cbfe6ee6091b3c) and new Gabriela's contact details, which we want to release with GAP 4.6 independently of the progress with other branches, so we publish it to be on the safe side.