Corrected example that depends on randomness.
Angel explained what happened:

The difference between the two output is encountered in the last entry.

In the both cases the last entry of the output represents a crossed product
Q(\xi)*(<a>\times <b>)
where
    \xi is a primitive 12-th root of unity,
    a and b have order 2,
the action is determined by
    \xi^a=\xi^5 and  \xi^b=\xi^7

The difference occurs in the twisting which in the first output is given by
    a^2=\xi^9, b^2 = 1, ba=\xi^9 ab.
and in the second output the twisting is given by
    a^2=\xi^3, b^2 = 1, ba=\xi^3 ab.

We can see that the two algebras are the same by observing that we have four
elections for the primitive root of unity, namely if \zeta=e^{\pi i}{6}, a
primitive 12-th then \xi can be either \zeta, \zeta^5, \zeta^7 or \zeta^{11}.
So if for the first algebra we use \xi_1=\zeta and for the second algebra we
use \xi_2=\zeta^{11}=\xi^{-12} then \xi_2^3=\xi_1^{-3}=\xi_1^9 and hence the
two algebras are the same.