Make div-alg examples independent of character ordering
Repaired the examples in the div-alg.xml file for the manual.  The
examples should now be free of character numbering issues.  In the
examples now the characters used are the unique ones up to Galois
conjugacy with kernels of a given size, so I added some lines that
will determine this character. So the output should be consistent
for the Schur index functions on all systems now.