Fix bug in ReducingCyclotomicAlgebra
The definition of the fields F1 and F2 have been modified.
F1=K^G1 and F2=K^G2 where K=F(E(m)) with F=A[2] and m=A[3].
The error consists in assuming that m coincides with the conductor of K.
G1 and G2 are expressed in terms of the information in A[4]
which in turn is expresed in terms of m. To define F1 and F2 properly
is necessarily to express the information of A[4] in terms of the conductor of K.
Another bug fixed in the return for the case where s1=s2=1, i.e. when the algebra is split.