Fix a bug in ReducingCyclotomicAlgebra: The order of the root of unity
This commit fixes a bug in the root of unity to be used for the cyclotomic
extension defining the coefficient field of one factor in the reducing
process of a cyclotomic algebra.

The function IsCyclotomicExtension has been modified so that the output is
a positive integer n if the input is of the form (K,F) with F an abelian
number field and K=F(E(n)). Otherwise the output is 0.

This is used in ReducingCyclotomicAlgebra for the production of the
factors of in the input as cyclotomic algebras. The factor Ai (i=1,2)
is a crossed product Fi*Gi, where Fi=F(E(ci)) and the info of the algebra
then takes the form [*,F,ci,*,*] where the last two entries are expressed
in terms of ci.