Fix KillingCocycle, speed up ReducingCyclotomicAlgebra
Fix a bug in KillingCocycle to prevent a non-equivalent
algebra output. This bug showed up with SG(672,622).

Improve checking that a cyclic algebra (K/F,a) with a
root of unity is split by verifying that a is the norm
of a root of unity simply comparing the order of a with
the order of the norm of a generator of the group of
roots of unity of K.