New version of IsDyadicSchurGroup
Fixes #103 reported by Thomas Breuer.

The fix provided by Allen Herman who wrote:

"The cause of the problem is that my earlier implementation of this
command does not recognize some dyadic Schur groups of type (QD,q)
correctly, this causes our current Wedderga to return a 2-local index
of 1 for these groups when it should be a 2. I believe the group
SmallGroup(272,15) that Thomas Breuer mentioned is the smallest
group with this issue.

The new implementation of the command follows the description of
the dyadic Schur groups of type (QD,q) in Riese and Schmid's 1996
paper very precisely, so I'm confident this will fix the problem."